Version 23.3.12
[Engine]
* Improve rounding by using proper floorf, ceilf and roundf functions when dealing with standard floating point numbers.
* VSync setting now applies immediately.
* Window handling changes were made which may fix and/or create some issues.

[Diggers]
* Changing VSync setting now applies immediately and added some additional settings.

[Source]
* README: Update some links and punctuation.
* AUDIO: Add 'final' class attribute.
* CMDLINE: Add 'final' class attribute.
* CODEC: Add 'final' class attribute.
* COLLECT: Add 'final' class attribute to all collector classes.
* CONLIB: Remove 'safe' function and do manual valid filename check.
* CONLIB: Use 'ceilf' instead of 'ceil'.
* CONSOLE: Use 'ceilf' instead of 'ceil'.
* CRYPT: Add 'final' class attribute.
* CVARLIB: Add callbacks for some cvars.
* CVARS: Add 'final' class attribute.
* DIR: Remove redundant 'safe' functions.
* DISPLAY: Add cvar callbacks to cache user settings in flags.
* DISPLAY: Cache user settings inside flags.
* DISPLAY: Set context robustness.
* DISPLAY: Set instance name in Linux.
* DISPLAY: Set window frame names in Mac and Linux.
* DISPLAY: Use more simple and readable function names.
* ERROR: Add 'final' class attribute.
* EVTMAIN: Add 'final' class attribute.
* EVTWIN: Add 'final' class attribute.
* FBOBASE: Add 'final' class attribute.
* FBOMAIN: Add 'final' class attribute.
* FBOMAIN: Use 'roundf' instead of 'round'.
* FONT: Move RoundFunc classes from Util class.
* FTF: Add 'final' class attribute.
* GLFW: Add 'final' class attribute.
* GLFW: Add functions for each hint and attribute.
* GLFW: Fix error handler multiple exceptions.
* GLFW: Simplify function creation using macros.
* IMGFMT: Add 'final' class attribute to all codec classes.
* INPUT: Add 'final' class attribute.
* LLASSET: Rename argument labels from 'Text' to 'String'.
* LLCORE: Rename argument labels from 'Text' to 'String'.
* LLDISPLAY: Update new simplified function names from GLFW.
* LLFILE: Add 'Executable' function to hide unused compiler warning.
* LLFILE: Use appropriate luautil function to check for safe file names.
* LLFONT: Rename argument labels from 'Text' to 'String'.
* LLUTIL: Rename argument labels from 'Text' to 'String'.
* LOG: Add 'final' class attribute.
* LUA: Add 'final' class attribute.
* OAL: Add 'final' class attribute.
* OGL: Add 'final' class attribute.
* OGL: Make VSync mode apply immediately.
* OGL: Remove unused mask value.
* PCMFMT: Add 'final' class attribute to all codec classes.
* SQL: Add 'final' class attribute.
* SQL: Add more functions to test error level for readability.
* SQL: Do not use log warning level on failed transaction.
* SQL: Hide routine "END TRANSACTION" attempts warnings.
* SSHOT: Add 'final' class attribute.
* STRING: Add 'final' class attribute.
* STRING: Add [[maybe_unused]] to unused (but useful) functions to remove MSVC compiler warnings.
* SYSCORE: Add 'final' class attribute.
* SYSUTIL: Add 'final' class attribute.
* TEXTURE: Use 'floorf' instead of 'floor'.
* UTIL: Fix MSVC 'unreachable code' compiler warning.
* UTIL: Move RoundFunc classes inside Font class.
* UTIL: Use 'ceilf' instead of 'ceil'.
* UTIL: Use 'floorf' instead of 'floor'.
* UTIL: Use 'roundf' instead of 'round'.
* DIGGERS: Change setup screen VSync option to accomodate new engine changes.
* MANUAL: Update reference manual to account for API changes.

+++ src/texture.hpp
@@ -464,11 +464,11 @@ BEGIN_MEMBERCLASS(Textures, Texture, ICHelperUnsafe),
         const GLfloat
           fTPSizeX = fTWidth+fPadX,
           fTPSizeY = fTHeight+fPadY,
-          fBTSizeX = floor(fBWidth/fTPSizeX)*fTPSizeX,
-          fBTSizeY = floor(fBHeight/fTPSizeY)*fTPSizeY;
+          fBTSizeX = floorf(fBWidth/fTPSizeX)*fTPSizeX,
+          fBTSizeY = floorf(fBHeight/fTPSizeY)*fTPSizeY;
         // Make sure theres enough memory allocated for each coord data
         tcList.reserve(static_cast<size_t>(
-          floor(fBTSizeX/fTPSizeX)*floor(fBTSizeY/fTPSizeY)));
+          floorf(fBTSizeX/fTPSizeX)*floorf(fBTSizeY/fTPSizeY)));
         // Image is reversed?
         if(bReversed)
           for(GLfloat fY = 0; fY < fBTSizeY; fY += fTPSizeY)
